requirements.txt
Django>=4.2
djangorestframework
psycopg2-binary
python-dotenv
Pillow
django-cors-headers
easyocr
opencv-python-headless